# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для работы с приёмами пациентов",
        "title": "ClinicHub API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/emk/{pat_id}": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EMK"
                ],
                "summary": "Get all EMK records by patient ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "pat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entities.Emk"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/medcard/{pat_id}": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalCard"
                ],
                "summary": "Get medical card by patient ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "pat_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entities.OneCMedicalCard"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalCard"
                ],
                "summary": "Update medical card by patient ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Patient ID",
                        "name": "pat_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Medical card update data",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.OneCMedicalCard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/onec/auth": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Syncs a batch of users (login + password) received from 1C into the internal auth system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1C"
                ],
                "summary": "Sync users from 1C",
                "parameters": [
                    {
                        "description": "List of users to sync",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AuthCreditionals"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/patients": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patients"
                ],
                "summary": "Get cached patient list from 1C with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 20, max: 100)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PatientListResponse"
                        }
                    }
                }
            }
        },
        "/webhook/onec/patients": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1C"
                ],
                "summary": "Webhook from 1C: patient list update",
                "parameters": [
                    {
                        "description": "Patient list update from 1C",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entities.PatientListUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/webhook/onec/receptions": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "1C"
                ],
                "summary": "Webhook from 1C: new receptions",
                "parameters": [
                    {
                        "description": "Receptions update",
                        "name": "update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Call"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ws/notification/register/{user_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websocket"
                ],
                "summary": "Подписаться на уведомления",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ws/notification/unregister/{user_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Websocket"
                ],
                "summary": "Отписаться от уведомлений",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entities.Analysis": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "entities.AttendingDoctor": {
            "type": "object",
            "properties": {
                "attachmentEnd": {
                    "type": "string"
                },
                "attachmentStart": {
                    "type": "string"
                },
                "clinic": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "policyOrCertNumber": {
                    "type": "string"
                },
                "specialization": {
                    "type": "string"
                }
            }
        },
        "entities.CallStatus": {
            "type": "string",
            "enum": [
                "completed",
                "process"
            ],
            "x-enum-comments": {
                "CallStatusCompleted": "исправлено: было \"compleated\"",
                "CallStatusWork": "исправлено: было \"proccess\""
            },
            "x-enum-descriptions": [
                "исправлено: было \"compleated\"",
                "исправлено: было \"proccess\""
            ],
            "x-enum-varnames": [
                "CallStatusCompleted",
                "CallStatusWork"
            ]
        },
        "entities.Certificate": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "number": {
                    "type": "string"
                }
            }
        },
        "entities.ClientRef": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "entities.DoctorData": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "specialization": {
                    "type": "string"
                }
            }
        },
        "entities.Emk": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/entities.Analysis"
                },
                "call_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "doctor": {
                    "$ref": "#/definitions/entities.DoctorData"
                },
                "flg": {
                    "$ref": "#/definitions/entities.Flg"
                },
                "id": {
                    "type": "integer"
                },
                "med_services": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "patient_id": {
                    "description": "связь с пациентом",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/entities.CallStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "entities.File": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "minioKey": {
                    "description": "\"minio_id\"",
                    "type": "string"
                }
            }
        },
        "entities.Flg": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "file": {
                    "$ref": "#/definitions/entities.File"
                },
                "file_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "number": {
                    "type": "string"
                },
                "organization": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "integer"
                },
                "result": {
                    "type": "string"
                }
            }
        },
        "entities.OneCMedicalCard": {
            "type": "object",
            "properties": {
                "additional_phone": {
                    "type": "string"
                },
                "address": {
                    "type": "string"
                },
                "age": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "certificate": {
                    "$ref": "#/definitions/entities.Certificate"
                },
                "display_name": {
                    "description": "Основные поля",
                    "type": "string"
                },
                "doctor": {
                    "$ref": "#/definitions/entities.AttendingDoctor"
                },
                "email": {
                    "type": "string"
                },
                "legal_representative": {
                    "description": "Вложенные структуры",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entities.ClientRef"
                        }
                    ]
                },
                "mobile_phone": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "policy": {
                    "$ref": "#/definitions/entities.Policy"
                },
                "relative": {
                    "$ref": "#/definitions/entities.Relative"
                },
                "snils": {
                    "type": "string"
                },
                "workplace": {
                    "type": "string"
                }
            }
        },
        "entities.OneCPatientListItem": {
            "type": "object",
            "properties": {
                "birthDate": {
                    "description": "в формате \"YYYY-MM-DD\"",
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "gender": {
                    "description": "true — мужской",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "patientID": {
                    "type": "string"
                }
            }
        },
        "entities.PatientListUpdate": {
            "type": "object",
            "properties": {
                "patients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.OneCPatientListItem"
                    }
                }
            }
        },
        "entities.Policy": {
            "type": "object",
            "properties": {
                "number": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "entities.Receptions": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "entities.Relative": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.AuthCreditionals": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handlers.OneCUser"
                    }
                }
            }
        },
...
        "handlers.OneCUser": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string",
                    "example": "doc1"
                },
                "password": {
                    "type": "string",
                    "example": "secret123"
                }
            }
        },
...
        "models.Call": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "analysis": {
                    "$ref": "#/definitions/entities.Analysis"
                },
                "call_id": {
                    "type": "string"
                },
                "doctor": {
                    "$ref": "#/definitions/entities.DoctorData"
                },
                "flg": {
                    "$ref": "#/definitions/entities.Flg"
                },
                "med_services": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "patients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.OneCPatientListItem"
                    }
                },
                "phone": {
                    "type": "string"
                },
                "receptions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.Receptions"
                    }
                },
                "status": {
                    "$ref": "#/definitions/entities.CallStatus"
                }
            }
        },
...
        "models.PatientListResponse": {
            "type": "object",
            "properties": {
                "patient": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entities.OneCPatientListItem"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer \u003cJWT\u003e\" to authenticate",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}